Update main.ts
if you run that script using a windows runner in github workflows, then it cannot create symlinks because in windows you need admin rights to do so. 
github does not provide the runner with such permisions.
therefore the symlink simply does not work.

It should- however - work when you directly set the environment variable to the path of th executable. 

I hope I did not miss, why those symlinks were set in the first place and that they are not needed somewhere else?